Feature ImageManager for fetching image without using an image view
ImageManager can fetch images in the same way as the extension on
`ImageView`. It works pretty much in the exact same way except it does
not use any pre or post configuration closures. If the URL already
exists in the cache, that resource will be used by default, however you
can opt-out of using cache by setting `useCache` to `false` at the call
site.

`Fetcher` is now equitable, it is used to remove fetchers when they are
done doing their job in `ImageManager`.

This should not affect the current implementation of Imaginary as
`ImageManager` is completely separate from the image view extensions
that we had (and still have).