Set `wantsLayer` to true before trying to add an animation
On macOS, layers are optional, so with the current implementation we
assume that it is always there which is a wrong assumption. To fix this
issue, we now set `wantsLayer` to `true` before adding the animation.

```
Setting the value of this property to true turns the view into a
layer-backed view—that is, the view uses a CALayer object to manage its
rendered content. Creating a layer-backed view implicitly causes the
entire view hierarchy under that view to become layer-backed. Thus, the
view and all of its subviews (including subviews of subviews) become
layer-backed. The default value of this property is false.
```