Use transition closure instead of setting the image
Setting the image again can cause the view to blink as it will render
the image on screen because the layer needs updating. By piggy backing
on `transitionClosure`, it will animate from one image to another, but
only if the old image exists. If the image view has no image, it will
simply assign the image to self as the previous implementation did.